+++ src/main/java/com/phantomwing/rusticdelight/loot/SquidsDropCalamariModifier.java
@@ -3,6 +3,7 @@ package com.phantomwing.rusticdelight.loot;
 import com.google.common.base.Suppliers;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import com.phantomwing.rusticdelight.Configuration;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -12,22 +13,21 @@ import net.minecraftforge.common.loot.IGlobalLootModifier;
 import net.minecraftforge.common.loot.LootModifier;
 import net.minecraftforge.registries.ForgeRegistries;
 import org.jetbrains.annotations.NotNull;
-
 import java.util.function.Supplier;
 
-public class AddItemModifier extends LootModifier {
-    public static final Supplier<Codec<AddItemModifier>> CODEC = Suppliers.memoize(()
+public class SquidsDropCalamariModifier extends LootModifier  {
+    public static final Supplier<Codec<SquidsDropCalamariModifier>> CODEC = Suppliers.memoize(()
             -> RecordCodecBuilder.create(inst -> codecStart(inst)
             .and(ForgeRegistries.ITEMS.getCodec().fieldOf("item").forGetter(m -> m.item))
             .and(Codec.INT.fieldOf("minAmount").forGetter(m -> m.minAmount))
             .and(Codec.INT.fieldOf("maxAmount").forGetter(m -> m.maxAmount))
-            .apply(inst, AddItemModifier::new)));
+            .apply(inst, SquidsDropCalamariModifier::new)));
 
     private final Item item;
     private final int minAmount;
     private final int maxAmount;
 
-    public AddItemModifier(LootItemCondition[] conditionsIn, Item item, int minAmount, int maxAmount) {
+    public SquidsDropCalamariModifier(LootItemCondition[] conditionsIn, Item item, int minAmount, int maxAmount) {
         super(conditionsIn);
 
         this.item = item;
@@ -37,6 +37,10 @@ public class AddItemModifier extends LootModifier {
 
     @Override
     protected @NotNull ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot, LootContext context) {
+        if (!Configuration.SQUIDS_DROP_CALAMARI.get()) {
+            return generatedLoot;
+        }
+
         for(LootItemCondition condition : this.conditions) {
             if(!condition.test(context)) {
                 return generatedLoot;