+++ src/main/java/com/phantomwing/rusticdelight/tags/ModTags.java
@@ -19,6 +19,7 @@ public class ModTags {
     // Item tags
     public static class Items {
         public static final TagKey<Item> COOKING_OIL_INGREDIENTS = tag("cooking_oil_ingredients");
+        public static final TagKey<Item> CALAMARI_ROLL_INGREDIENTS = tag("calamari_roll_ingredients");
 
         private static TagKey<Item> tag(String name) {
             return ItemTags.create(new ResourceLocation(RusticDelight.MOD_ID, name));